implementando o controller WorkController e algumas correcoes

+++ src/main/java/br/com/github/renato28/composicaospring/repositories/HourContractRepository.java
@@ -3,5 +3,5 @@ package br.com.github.renato28.composicaospring.repositories;
 import br.com.github.renato28.composicaospring.entities.HourContract;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface HourContractRepository extends JpaRepository<HourContract, Integer> {
+public interface HourContractRepository extends JpaRepository<HourContract, Long> {
 }
